fix: /EHsc on MSVC + Scheme signedness; bump to 0.3.2
Two fixes that make the crate build again with current toolchains
(MSVC 14.44 / VS 2022 17.14, bindgen 0.72):

The cmake build passes cc-derived base flags that REPLACE MSVC's defaults,
dropping `/EHsc` -- and MSVC 14.44's STL escalates the resulting C4530
("C++ exception handler used, but unwind semantics are not enabled") to an
error via C2220, first surfacing in `regression/common/shape_utils.cpp`
(compiled despite `NO_REGRESSION=1`). OpenSubdiv's STL usage always assumed
unwind semantics; older MSVC merely warned. `build.rs` now adds
`cxxflag("/EHsc")` on `target_env = "msvc"`.

`TopologyRefiner::options()` converted `_subdivType` with a bare
`.try_into()`. bindgen's signedness for C++ enum fields varies by platform
and version -- MSVC enums come through as `i32` -- while `Scheme` is
`repr(u32)` with `TryFromPrimitive`, so the conversion no longer resolved.
Route through `u32` explicitly; invalid discriminants still fail in
`try_into`.

Verified downstream (akatela, Windows, MSVC 14.44): OpenSubdiv compiles
clean and the full akatela-io suite passes with the `opensubdiv` feature
enabled, 109/109 -- the first green build of this feature on that toolchain.

PUBLISHING NOTE: master also carries API changes relative to the published
0.3.1 -- `refine_adaptive` takes `Option<&[Index]>` where 0.3.1 took a
slice, and `TopologyDescriptor::{creases, corners}` are consuming builders
returning `Result` where 0.3.1 mutated in place. Under the cargo semver
convention 0.3.1 -> 0.3.2 is a compatible upgrade, so `^0.3` users calling
those APIs will break on `cargo update`; 0.4.0 would signal the break.
Versioned 0.3.2 per maintainer decision.

+++ opensubdiv-petite/src/far/topology_refiner.rs
@@ -72,7 +72,12 @@ impl TopologyRefiner {
     pub fn options(&self) -> TopologyRefinerOptions {
         let options = unsafe { &(*self.0)._subdivOptions };
         TopologyRefinerOptions {
-            scheme: unsafe { (*self.0)._subdivType }
+            // Cast: bindgen's signedness for the C++ enum field varies by
+            // platform/version (MSVC enums come through as `i32`), while
+            // `Scheme` is `repr(u32)` with `TryFromPrimitive` -- so route
+            // through `u32` explicitly. Invalid discriminants still fail in
+            // `try_into`.
+            scheme: (unsafe { (*self.0)._subdivType } as u32)
                 .try_into()
                 .expect("invalid subdivision scheme from C++"),
             boundary_interpolation: if options._vtxBoundInterp
